Remove phrases always from the user database when typing Alt+Number
It should not matter whether the phrase selected with Alt+Number is a
system phrase or user defined phrase, in both cases it should be
removed from the user database to score down the likelyhood of this
phrase being suggested again.

The user should not need to think about whether this is a system
phrase or a user defined phrase, selecting a phrase with Alt+Number
just means: “I do not want this to be suggested here, reduce the
score of this”.

Also, make the function remove_phrase() use keyword arguments
for input_phrase and phrase instead of passing a complete
database row where most of the columns are ignored anyway,
this improves readability a bit, I think.

+++ ibus-typing-booster/engine/tabsqlitedb.py
@@ -615,19 +615,24 @@ CREATE TABLE phrases (id INTEGER PRIMARY KEY AUTOINCREMENT,                mlen 
                                database='mudb'),
             mudb.keys())
 
-    def remove_phrase (self,phrase,database='user_db'):
+    def remove_phrase (self, input_phrase='', phrase='', database='user_db'):
         '''
-        Remove phrase from database.
-        phrase should be a tuple like a row in the database, i.e.
-        like the result lines of a "select * from phrases;"
-        Like (id, mlen,clen,input_phrase,phrase,freq,user_freq)
+        Remove all rows matching “input_phrase” and “phrase” from database.
+        Or, if “input_phrase” is “None”, remove all rows matching “phrase”
+        no matter for what input phrase from the database.
         '''
-        id, mlen, clen, input_phrase, phrase, freq, user_freq = phrase
-
-        delete_sqlstr = '''
-        DELETE FROM %(database)s.phrases
-        WHERE input_phrase = :input_phrase AND phrase = :phrase
-        ;''' %{'database': database}
+        if not phrase:
+            return
+        if input_phrase:
+            delete_sqlstr = '''
+            DELETE FROM %(database)s.phrases
+            WHERE input_phrase = :input_phrase AND phrase = :phrase
+            ;''' %{'database': database}
+        else:
+            delete_sqlstr = '''
+            DELETE FROM %(database)s.phrases
+            WHERE phrase = :phrase
+            ;''' %{'database': database}
         delete_sqlargs = {'input_phrase': input_phrase, 'phrase': phrase}
         self.db.execute(delete_sqlstr, delete_sqlargs)
         self.db.commit()
